OMETiffReader: set currentID on the first match if not set
This fixes a bug which appeared for partial multi-file OME-TIFF filesets with
no OME.UUID element in the metadata. TiffData elements pointing to missing
files would be incorrectly assigned to the initialized file.

The proposed change ensures the uninitialized currentUUID variable gets
assigned as soon as the reader maps a TiffData element to the current file.
All subsequent iterations of the mapping loops can then use this UUID to check
for missing files/inconsistent mappings.