Olympus vsi: fix exposure time handling
Exposure times can be recorded with no tag prefix or with the "Microscope " tag prefix.
Some datasets have just one or the other, some have both.
The times with no prefix are preferred if present, but the "Microscope " ones are now
used if that is all that is available.

Fixes #3642.