Fix min/max calculation for RGB data
This fixes an arithmetic error which prevented correct min/max values
from being reported for RGB data. The length supplied to updateMinMax(...)
is expected to be the total number of bytes, including any RGB channels.
updateMinMax(...) later performs calculations on the length to determine the
number of pixels, which are incorrect if the RGB channels are not included.

Non-RGB data (getRGBChannelCount() == 1) is unaffected, so anything that
includes ChannelSeparator as well as MinMaxCalculator in the reader stack
should not be impacted by this change.