Fix NPEs when a no-op file pattern is encountered
If the helper reader performs file grouping, then `externals` will be null
and `noStitch` will be set to `true`. This cleans up a few overlooked
places where `externals` was assumed to be non-null. It also logs
an INFO message if the file pattern is discarded in favor of grouping
by the helper reader.