vsi: slightly rework tile decoding to fail faster on invalid tiles
This reads the expected number of compressed bytes into an array,
then attempts to decompress the array. This will fail faster on invalid
tiles than passing the entire .ets file to a codec, which may attempt
to read the entire file in search of the start or end of the compressed
block.