Add API to configure the value of undefined pixels
Fixes #3800.

This generalizes the work in https://github.com/glencoesoftware/bioformats2raw/pull/65
to allow a fill value to be set on all readers. Unless otherwise specified, the fill value
will be 0, so the default behavior should be unchanged.

ZeissCZIReader is currently the only reader that overrides the default fill value.
This is an extension of #3795.

showinf and bfconvert also have a new `-fill byte` option that uses the new API.