Add 'unconfigured' tool to run ReportEnabledStatus
unconfigured and unconfigured.bat should end up in bftools.zip.
Adding bio-formats-testing-framework.jar to the bftools directory
should allow both commands to be run, with syntax:

$ unconfigured /path/to/data /path/to/configuration [--enabled]

If the "--enabled" option is omitted, a list of files that are configured as
"test = false" with no explanatory comment will be printed.
If the "--enabled" option is included, a list of files that are configured as
"test = true" will be printed.
Omitting either directory or including extra arguments should result in an error.