Add -extra-metadata option for bfconvert
Fiddled with the API a bit to make this work in a more extensible way.
Supplying an extra metadata location is no longer tied to DicomWriter
specifically, but is in an extra interface that can be implemented
by writers that support this feature.

Considered implementing bfconvert connectivity via an option in DicomWriter,
instead of a new command line argument in bfconvert. Either way would work,
but the approach here would allow us to implement a similar feature in
other writers later on (if we choose to do so).