Add configurable tag resolution strategy
Each tag contained in a JSON file may now optionally contain a
"ResolutionStrategy" property set to "IGNORE", "REPLACE", or "APPEND".
IGNORE means that the tag will be ignored if the same tag code has been defined already.
REPLACE means that the tag will be used to replace any existing tag with the same code.
APPEND means that if there is an existing tag with the same code, the current tag's
value will be appended to the pre-existing tag's value array.