Remove assertion that would never fail
MCU sizes will be read as long as SOF0 is read. SOF0 is assumed to be read since it contains the dimensions as well. We should either fail on not finding a SOF0 or finding multiple of SOF0 tags. So if there should be an assertion, it should be "exactly one SOF0 found", that would be more sensible.