SimplePCI: restrict "Position_Z" check so that "*Position_Z_Fine" is excluded
Fixes #4133.

As noted there, .cxd files produced by newer versions of HCImage include
both a "Position_Z" and "Position_Z_Fine" value. This can result in the
"uniqueZ" list being larger than the number of detected planes. That
causes line 381 to set sizeT to 0, which means that the following "while"
loop runs until sizeZ and sizeT overflow to negative values.

This intentionally does not check that the name equals "Position_Z",
as some older files add a prefix to this name.