CV7000: handle NullPointerException when resetting allFiles in close()
While the reader should initialize this field as an empty ArrayList, there are
scenarios where allFiles will be null when calling close().
A use case would be a reader initialized from a memo file created with a
previous release of Bio-Formats since this field was formerly a string array
set to null by default.