Update SlideBook7Reader.java
There was a problem setting PhysicalSizeZ is the value was negative. Changed to use the absolute value of Interplane Spacing
Uploaded a test file to Zenodo:
https://zenodo.org/records/16355849
Old version will not display PhysicalSizeZ  in the OME Metadata window. New version will show it with a value of 100 for this test case.